requirements.txt
streamlit>=1.39.0
langchain>=0.2.11
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langchain-google-genai>=1.0.7
google-generativeai>=0.7.2
chromadb>=0.5.5
langchain-chroma>=0.1.0
sentence-transformers>=3.0.1
python-dotenv>=1.0.1
pypdf>=4.3.1
pdfminer.six>=20221105
docx2txt>=0.8
pandas>=2.2.3
pytest>=8.3.2
watchdog>=4.0.0
google-api-python-client>=2.147.0
google-auth>=2.33.0
